pyproject.toml
[tool.poetry]
name = "aws_jenkins"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [
   { include = "tvm_ci", from="python" },
]

[tool.poetry.dependencies]
python = "^3.7"
PyYAML = "^5.4.1"
requests = "^2.25.1"
boto3 = "1.17.67"  # TODO switch back to semver
jenkins-job-builder = {git = "https://github.com/areusch/jenkins-job-builder", rev = "master"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
